🌊 verifications

+++ UI/RockOutcrop.cpp
@@ -41,9 +41,11 @@ RockOutcrop::RockOutcrop(QWidget *parent) :
         newDir.mkpath(".");
     }
 
-
-    QDir resDir(QDir(rootDir).filePath("resources"));
-    resDir.removeRecursively();
+    if(!loadPreviousResults)
+    {
+        QDir resDir(QDir(rootDir).filePath("resources"));
+        resDir.removeRecursively();
+    }
 
     copyDir(QDir(qApp->applicationDirPath()).filePath("resources"),QDir(rootDir).filePath("resources"),true);
 
@@ -290,7 +292,7 @@ RockOutcrop::RockOutcrop(QWidget *parent) :
     {
         elementModel->clear();
         elementModel->refresh();
-        //loadFromJson(); // commenting this for frank,
+        if(loadPreviousResults) loadFromJson(); // commenting this for frank,
         //because he doesn't like previous analysis loaded by default
 
         if(ui->tableView->m_sqlModel->rowCount()<1)
@@ -411,6 +413,18 @@ void RockOutcrop::loadFromJson()
     QJsonDocument indoc = QJsonDocument::fromJson(in.toUtf8());
     QJsonObject inobj = indoc.object();
 
+
+
+
+    QJsonObject basicSettings = inobj["basicSettings"].toObject();
+    QString groundMotion = basicSettings["groundMotion"].toString();
+    theTabManager->updateGMPath(groundMotion);
+    QString OpenSeesPath = basicSettings["OpenSeesPath"].toString();
+    theTabManager->updateOpenSeesPath(OpenSeesPath);
+    QString slopex1 = QString::number(basicSettings["slopex1"].toDouble(), 'g', 16);
+    QString slopex2 = QString::number(basicSettings["slopex2"].toDouble(), 'g', 16);
+
+
     QJsonArray soilLayers = inobj["soilProfile"].toObject()["soilLayers"].toArray();
     QJsonArray materials = inobj["materials"].toArray();
     for (int i=soilLayers.size()-1; i>=0; i--)
@@ -439,9 +453,12 @@ void RockOutcrop::loadFromJson()
             if (color=="")
                 color = QColor::fromRgb(QRandomGenerator::global()->generate()).name();
             valueList << name << thickness << density << vs << material << eSize << color;
-            ui->tableView->insertAtSilent(valueList,0);
+            //ui->tableView->insertAtSilent(valueList,0);
+            ui->tableView->insertAt(valueList,0);
 
         }
+
+
 
         theTabManager->updateLayerTab(l,mat);
 
@@ -449,18 +466,20 @@ void RockOutcrop::loadFromJson()
 
 
 
-    QJsonObject basicSettings = inobj["basicSettings"].toObject();
-    QString groundMotion = basicSettings["groundMotion"].toString();
-    theTabManager->updateGMPath(groundMotion);
-    QString OpenSeesPath = basicSettings["OpenSeesPath"].toString();
-    theTabManager->updateOpenSeesPath(OpenSeesPath);
+
+
+
 
     ui->gwtEdit->setText(QString::number(basicSettings["groundWaterTable"].toDouble()));
 
     ui->totalLayerLineEdit->setText(QString::number(soilLayers.size()));
 
+    theTabManager->updateConfigureTabFromOutside(slopex1, slopex2);
+
 
     ui->reBtn->click();
+
+
 
 }
 
@@ -1006,31 +1025,31 @@ void RockOutcrop::on_reBtn_clicked()
             double eSize = atof(pars[0].c_str());
             int id = i;
 
-            int DrInd=0, hPermInd=0, vPermInd=0, uBulkInd=0;
+            int DrInd=0, hPermInd=0, vPermInd=0, uBulkInd=0, voidInd=0;
             if(list.at(MATERIAL-2).toString()=="Elastic")
             {
-                DrInd = 4; hPermInd = 6; vPermInd = 7; uBulkInd = 10;
+                DrInd = 4; hPermInd = 6; vPermInd = 7; uBulkInd = 10; voidInd = 5;
             }else if(list.at(MATERIAL-2).toString()=="PM4Sand")
             {
-                DrInd = 1; hPermInd = 25; vPermInd = 26; uBulkInd = 27;
+                DrInd = 1; hPermInd = 25; vPermInd = 26; uBulkInd = 27; voidInd = 28;
             }else if(list.at(MATERIAL-2).toString()=="PM4Silt")
             {
-                DrInd = 1; hPermInd = 27; vPermInd = 28; uBulkInd = 29;
+                DrInd = 1; hPermInd = 27; vPermInd = 28; uBulkInd = 29; voidInd = 30;
             }else if(list.at(MATERIAL-2).toString()=="PIMY")
             {//TODO: double check
-                DrInd = 1; hPermInd = 12; vPermInd = 13; uBulkInd = 14;
+                DrInd = 1; hPermInd = 12; vPermInd = 13; uBulkInd = 14; voidInd = 15;
             }else if(list.at(MATERIAL-2).toString()=="PDMY")
             {//TODO: double check
-                DrInd = 1; hPermInd = 24; vPermInd = 25; uBulkInd = 26;
+                DrInd = 1; hPermInd = 24; vPermInd = 25; uBulkInd = 26; voidInd = 27;
             }else if(list.at(MATERIAL-2).toString()=="PDMY02")
             {//TODO: double check
-                DrInd = 1; hPermInd = 26; vPermInd = 27; uBulkInd = 28;
+                DrInd = 1; hPermInd = 26; vPermInd = 27; uBulkInd = 28; voidInd = 29;
             }else if(list.at(MATERIAL-2).toString()=="ManzariDafalias")
             {//TODO: double check
-                DrInd = 1; hPermInd = 20; vPermInd = 21; uBulkInd = 22;
+                DrInd = 1; hPermInd = 20; vPermInd = 21; uBulkInd = 22; voidInd = 23;
             }else if(list.at(MATERIAL-2).toString()=="J2Bounding")
             {//TODO: double check
-                DrInd = 1; hPermInd = 10; vPermInd = 11; uBulkInd = 12;
+                DrInd = 1; hPermInd = 10; vPermInd = 11; uBulkInd = 12; voidInd = 13;
             }
 
 
@@ -1055,6 +1074,7 @@ void RockOutcrop::on_reBtn_clicked()
                 {"hPerm",thisFEMList.at(hPermInd).toDouble()},
                 {"vPerm",thisFEMList.at(vPermInd).toDouble()},
                 {"uBulk",thisFEMList.at(uBulkInd).toDouble()},
+                {"void",thisFEMList.at(voidInd).toDouble()},
             };
             material =  createMaterial(i+1, list.at(MATERIAL-2).toString().toStdString(),list.at(FEM-2).toString().toStdString());
             materials.push_back(material);
